tuning engine to support shuffle list feature and stop within 5 seconds.

+++ ElectricityFactory/Planning.cs
@@ -2,10 +2,10 @@ namespace ElectricityFactory
 {
     public class Planning
     {
-        public int TotalCoalAmount { get; set; } 
-        public int AveragePrice { get; set; } 
-        public int AverageQnetThreshold { get; set; } 
-        public int AverageVadThreshold { get; set; } 
+        public float TotalCoalAmount { get; set; } 
+        public float AveragePrice { get; set; } 
+        public float AverageQnetThreshold { get; set; } 
+        public float AverageVadThreshold { get; set; } 
         public float AverageSadThreshold { get; set; } 
     }
 }
